refactor: add interface for registration service and manage registration controller
