+++ src/main/java/com/ipz/mba/controllers/RegistrationsController.java
@@ -3,6 +3,7 @@ package com.ipz.mba.controllers;
 import com.ipz.mba.models.ClientDataRegistration;
 import com.ipz.mba.security.jwt.JWTUtil;
 import com.ipz.mba.services.RegistrationService;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -10,7 +11,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.Map;
-
+@Slf4j
 @RestController
 @RequestMapping("/register")
 public class RegistrationsController {
@@ -27,9 +28,8 @@ public class RegistrationsController {
     public Map<String, String> register(@RequestBody ClientDataRegistration clientData) {
         String jwtToken;
         try {
-            System.out.println("LOG: " + clientData);
+            log.info("LOG: " + clientData);
             registrationService.saveData(clientData);
-
             jwtToken = jwtUtil.generateToken(clientData.getPhoneNumber() != null ?
                     clientData.getPhoneNumber() : clientData.getIpn());
         } catch (Exception ex) {
